Workaround for webkit position:absolute bug
The menu bar used to be composed of a number of absolute positioned divs
inside a relative positioned wrapper to maintain the z-index I wanted.
I've discovered and reported a bug where adding any <form> to a page
breaks absolute positioned elements onLoad.

As a workaround, the menu bar was reworked to use other positioning
styles.